Assertion failure when requesting that storage be deleted for an origin without storage. https://bugs.webkit.org/show_bug.cgi?id=60709
Reviewed by Darin Adler.

* storage/StorageTracker.cpp:
(WebCore::StorageTracker::syncDeleteOrigin):
If the path to the storage location for the givin origin is empty, return early instead of
asserting.



Canonical link: https://commits.webkit.org/76028@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@86371 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/storage/StorageTracker.cpp
@@ -452,10 +452,13 @@ void StorageTracker::syncDeleteOrigin(const String& originIdentifier)
     openTrackerDatabase(false);
     if (!m_database.isOpen())
         return;
-    
-    String path = databasePathForOrigin(originIdentifier);
 
-    ASSERT(!path.isEmpty());
+    String path = databasePathForOrigin(originIdentifier);
+    if (path.isEmpty()) {
+        // It is possible to get a request from the API to delete the storage for an origin that
+        // has no such storage.
+        return;
+    }
     
     SQLiteStatement deleteStatement(m_database, "DELETE FROM Origins where origin=?");
     if (deleteStatement.prepare() != SQLResultOk) {
